+++ Extra/Settings/SettingGui.js
@@ -134,6 +134,10 @@ class SettingGui {
         this.addSlider("Head Scale", "headScale", this.defaultSettings.headScale || 8, 2, 15)
         this.addSlider("Icon Scale", "iconScale", this.defaultSettings.iconScale || 10, 2, 15)
 
+        this.addColorSelector("Border Color", "borderColor", this.defaultSettings.borderColor)
+        this.addColorSelector("Map Color", "mapBackgroundColor", this.defaultSettings.mapBackgroundColor)
+        this.addColorSelector("Extra Info Color", "extraInfoBackroundColor", this.defaultSettings.extraInfoBackroundColor)
+
         this.addCategory("Secret info Settings")
 
         this.addDropdown("Score info under map style", {
@@ -362,13 +366,103 @@ class SettingGui {
      * @param {String} setting internal name of the setting to control
      * @param {String} defau Default value
      */
-    addString(label, setting, defau) {
+     addString(label, setting, defau) {
         let textBox = this.addSidebarElement(new TextBox().setText(this.defaultSettings[setting] ?? defau).addEvent(new SoopyContentChangeEvent().setHandler((val, prev, cancelFun) => {
             this.changed(setting, val)
         })), 0.625, 0.2, 0.05)
 
         return [textBox, this.addSidebarElement(new SoopyTextElement().setText("§0" + label).setMaxTextScale(2), 0.1, 0.35)]
     }
+
+    /**
+     * 
+     * @param {String} label The text to go to the left of the slider
+     * @param {String} setting internal name of the setting to control
+     * @param {Object[]} defau Default value
+     */
+    addColorSelector(label, setting, defau) {
+        // Slider 0
+        let slider0 = new Slider().setValue(this.defaultSettings[setting][0] ?? defau).setMin(0).setMax(255).addEvent(new SoopyContentChangeEvent().setHandler((val, prev, cancelFun) => {
+            this.changedArr(setting, 0, Math.round(val))
+            numberT0.setText(Math.round(val).toString())
+        }))
+
+        let numberT0 = new NumberTextBox().setText((this.defaultSettings[setting][0] ?? defau).toString())
+
+        numberT0.isNumber = isNumber
+
+        numberT0.text.addEvent(new SoopyContentChangeEvent().setHandler((val, prev, cancelFun) => {
+            if (!val) return
+
+            this.changedArr(setting, 0, parseInt(val))
+            slider0.setValue(parseInt(val))
+        }))
+        this.addSidebarElement(slider0, 0.26, 0.09, 0.05).setLore(["red"])
+        this.addSidebarElement(numberT0, 0.35, 0.08, 0.05).setLore(["red"])
+        
+        // Slider 1
+        let slider1 = new Slider().setValue(this.defaultSettings[setting][1] ?? defau).setMin(0).setMax(255).addEvent(new SoopyContentChangeEvent().setHandler((val, prev, cancelFun) => {
+            this.changedArr(setting, 1, Math.round(val))
+            numberT1.setText(Math.round(val).toString())
+        }))
+
+        let numberT1 = new NumberTextBox().setText((this.defaultSettings[setting][1] ?? defau).toString())
+
+        numberT1.isNumber = isNumber
+
+        numberT1.text.addEvent(new SoopyContentChangeEvent().setHandler((val, prev, cancelFun) => {
+            if (!val) return
+
+            this.changedArr(setting, 1, parseInt(val))
+            slider1.setValue(parseInt(val))
+        }))
+
+        this.addSidebarElement(slider1, 0.44, 0.09,  0.05).setLore(["green"])
+        this.addSidebarElement(numberT1, 0.53, 0.08, 0.05).setLore(["green"])
+
+        // Slider 2
+        let slider2 = new Slider().setValue(this.defaultSettings[setting][2] ?? defau).setMin(0).setMax(255).addEvent(new SoopyContentChangeEvent().setHandler((val, prev, cancelFun) => {
+            this.changedArr(setting, 2, Math.round(val))
+            numberT2.setText(Math.round(val).toString())
+        }))
+
+        let numberT2 = new NumberTextBox().setText((this.defaultSettings[setting][2] ?? defau).toString())
+
+        numberT2.isNumber = isNumber
+
+        numberT2.text.addEvent(new SoopyContentChangeEvent().setHandler((val, prev, cancelFun) => {
+            if (!val) return
+
+            this.changedArr(setting, 2, parseInt(val))
+            slider2.setValue(parseInt(val))
+        }))
+
+        this.addSidebarElement(slider2, 0.62, 0.09,  0.05).setLore(["blue"])
+        this.addSidebarElement(numberT2, 0.71, 0.08, 0.05).setLore(["blue"])
+
+        // Slider 3
+        let slider3 = new Slider().setValue(this.defaultSettings[setting][3] ?? defau).setMin(0).setMax(255).addEvent(new SoopyContentChangeEvent().setHandler((val, prev, cancelFun) => {
+            this.changedArr(setting, 3, Math.round(val))
+            numberT3.setText(Math.round(val).toString())
+        }))
+
+        let numberT3 = new NumberTextBox().setText((this.defaultSettings[setting][3] ?? defau).toString())
+
+        numberT3.isNumber = isNumber
+
+        numberT3.text.addEvent(new SoopyContentChangeEvent().setHandler((val, prev, cancelFun) => {
+            if (!val) return
+
+            this.changedArr(setting, 3, parseInt(val))
+            slider3.setValue(parseInt(val))
+        }))
+
+        this.addSidebarElement(slider3, 0.81, 0.09,  0.05).setLore(["opacity"])
+        this.addSidebarElement(numberT3, 0.9, 0.08, 0.05).setLore(["opacity"])
+
+
+        return [[slider0, numberT0], this.addSidebarElement(new SoopyTextElement().setText("§0" + label).setMaxTextScale(2), 0.1, 0.15)]
+    } 
 
     /**
      * 
@@ -416,6 +510,13 @@ class SettingGui {
     }
 
 
+    /**
+     * @param {Object} elm The element being added
+     * @param {Number} x the x pos
+     * @param {Number} width The width of the elm
+     * @param {Number} height The height of the elm
+     * @returns {Object} elm
+     */
     addSidebarElement(elm = null, x = 0.1, width = 0.8, height = 0.1) {
         if (elm) this.mainpage.addChild(elm.setLocation(x, this.y + 0.05 - height / 2, width, height))
         if (x === 0.1) this.y += 0.1
@@ -543,4 +644,10 @@ function findKey(callback = () => { }) {
 
         new Notification("§cUnable to find api key", [])
     }).start()
+}
+
+function isNumber(val) {
+    if (val.includes(".")) return false
+    val = "" + val; //coerce num to be a string
+    return !isNaN(val) && !isNaN(parseInt(val));
 }